LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/container-selection-unknown-features-expected.txt

Green
Green
Green
Green

FAIL width query with (foo: bar) assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
FAIL width query with foo(bar) assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
PASS style query with (foo: bar)
